+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Features
+
+- Add `proguardUuid` option to `SentryOptions` ([#436](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/436))
+  - This will propagate the `proguardUuid` value to Sentry Android
+
 ## 0.17.1
 
 ### Fixes